# Use STArray as underlying blockRam representation