--- a/changelog/2021-07-21T10_34_48+02_00_asyncrom_x_exception
+++ b/changelog/2021-07-21T10_34_48+02_00_asyncrom_x_exception
@@ -1,0 +1,1 @@
+CHANGED: `asyncRom` now throws `XException` for out-of-bounds addressing, no longer aborting simulation