Improves simulation performance.

Before:
```
benchmarking RAMs/blockRam# (100% writes)
time                 3.225 ms   (3.184 ms .. 3.274 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 3.239 ms   (3.215 ms .. 3.267 ms)
std dev              87.20 μs   (73.02 μs .. 106.2 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.393e7    (1.393e7 .. 1.393e7)
  y                  3465.257   (2584.017 .. 4323.712)
variance introduced by outliers: 12% (moderately inflated)

benchmarking RAMs/blockRam# (0% writes)
time                 1.627 ms   (1.614 ms .. 1.639 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.627 ms   (1.622 ms .. 1.637 ms)
std dev              24.32 μs   (18.10 μs .. 35.70 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              8033976.000 (8033976.000 .. 8033976.000)
  y                  2224.000   (2224.000 .. 2224.000)

benchmarking RAMs/blockRamFile# (100% writes)
time                 2.934 ms   (2.898 ms .. 2.976 ms)
                     0.997 R²   (0.994 R² .. 0.999 R²)
mean                 2.994 ms   (2.964 ms .. 3.039 ms)
std dev              119.0 μs   (91.83 μs .. 158.2 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.087e7    (1.087e7 .. 1.087e7)
  y                  3453.379   (2642.869 .. 4251.116)
variance introduced by outliers: 23% (moderately inflated)

benchmarking RAMs/blockRamFile# (0% writes)
time                 1.409 ms   (1.387 ms .. 1.429 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 1.424 ms   (1.410 ms .. 1.438 ms)
std dev              46.97 μs   (36.69 μs .. 64.16 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              4980031.940 (4980031.800 .. 4980032.000)
  y                  2197.868   (2192.000 .. 2210.828)
variance introduced by outliers: 21% (moderately inflated)
```

After:
```
benchmarking RAMs/blockRam# (100% writes)
time                 1.201 ms   (1.185 ms .. 1.218 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 1.202 ms   (1.193 ms .. 1.213 ms)
std dev              34.59 μs   (26.87 μs .. 46.99 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              9162800.251 (9162791.332 .. 9162808.822)
  y                  3490.396   (2822.287 .. 4237.928)
variance introduced by outliers: 17% (moderately inflated)

benchmarking RAMs/blockRam# (0% writes)
time                 1.083 ms   (1.069 ms .. 1.098 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 1.052 ms   (1.043 ms .. 1.062 ms)
std dev              30.95 μs   (26.18 μs .. 38.64 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              8897232.000 (8897232.000 .. 8897232.000)
  y                  2224.000   (2224.000 .. 2224.000)
variance introduced by outliers: 18% (moderately inflated)

benchmarking RAMs/blockRamFile# (100% writes)
time                 1.118 ms   (1.105 ms .. 1.130 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 1.098 ms   (1.090 ms .. 1.108 ms)
std dev              30.07 μs   (22.92 μs .. 45.09 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              8233407.935 (8233407.786 .. 8233408.000)
  y                  2200.027   (2192.000 .. 2218.333)
variance introduced by outliers: 16% (moderately inflated)

benchmarking RAMs/blockRamFile# (0% writes)
time                 985.4 μs   (958.4 μs .. 1.009 ms)
                     0.997 R²   (0.996 R² .. 0.999 R²)
mean                 949.0 μs   (939.5 μs .. 959.4 μs)
std dev              33.01 μs   (25.21 μs .. 42.66 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              7967840.865 (7967834.549 .. 7967846.614)
  y                  3383.609   (2808.387 .. 3988.479)
variance introduced by outliers: 24% (moderately inflated)
```